+++ src/main/java/roomescape/application/JwtPayload.java
@@ -1,0 +1,6 @@
+package roomescape.application;
+
+import roomescape.domain.Role;
+
+public record JwtPayload(Long memberId, String name, Role role) {
+}
