+++ src/main/java/roomescape/presentation/interceptor/AdminInterceptor.java
@@ -1,44 +1,34 @@
 package roomescape.presentation.interceptor;
 
-import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import java.util.Arrays;
 import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.HandlerInterceptor;
-import roomescape.application.JwtProvider;
-import roomescape.application.JwtProvider.JwtPayload;
+import roomescape.application.JwtPayload;
 import roomescape.common.exception.ForbiddenException;
-import roomescape.common.exception.UnauthorizedException;
 import roomescape.domain.Role;
+import roomescape.infrastructure.security.JwtProvider;
+import roomescape.presentation.JwtTokenExtractor;
 
 @Component
 public class AdminInterceptor implements HandlerInterceptor {
 
+    private final JwtTokenExtractor jwtTokenExtractor;
     private final JwtProvider jwtProvider;
 
-    public AdminInterceptor(JwtProvider jwtProvider) {
+    public AdminInterceptor(JwtTokenExtractor jwtTokenExtractor, JwtProvider jwtProvider) {
+        this.jwtTokenExtractor = jwtTokenExtractor;
         this.jwtProvider = jwtProvider;
     }
 
     @Override
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler)
             throws Exception {
-        Cookie tokenCookie = getTokenCookie(request);
-        JwtPayload jwtPayload = jwtProvider.extractPayload(tokenCookie.getValue());
+        String jwtToken = jwtTokenExtractor.extract(request);
+        JwtPayload jwtPayload = jwtProvider.extractPayload(jwtToken);
         if (jwtPayload.role() != Role.ADMIN) {
             throw new ForbiddenException("접근 권한이 없습니다.");
         }
         return true;
-    }
-
-    private static Cookie getTokenCookie(HttpServletRequest request) {
-        if (request.getCookies() == null) {
-            throw new UnauthorizedException("인증 쿠키값이 존재하지 않습니다.");
-        }
-        return Arrays.stream(request.getCookies())
-                .filter(each -> each.getName().equals("token"))
-                .findFirst()
-                .orElseThrow(() -> new UnauthorizedException("인증 쿠키값이 존재하지 않습니다."));
     }
 }
